Campus Edit Functionality

+++ app/components/Campus.jsx
@@ -5,7 +5,10 @@ import axios from 'axios';
 export default class Campus extends Component{
     constructor(){
         super();
-        this.state={campuses:[]};
+        this.state={
+            campuses:[],
+            new:true
+        };
 
         this.handleClick=this.handleClick.bind(this);
     }
@@ -17,39 +20,38 @@ export default class Campus extends Component{
     }
 
     handleClick(e){
-        console.log("clickTarg",e.target.id);
-        axios({
-            method: 'delete',
-            url: '/api/campuses/'+e.target.id
-        })
-        .then(res =>{
-            // console.log('change state',res);
-            //cannot push to the same path must modify state
-            // this.props.history.push('/campuses');
-            // this.setState({ campuses });
-
-
-            // could not do another query on the server side
-            // please see api
-            axios.get('/api/campuses')
-            .then(res => res.data)
-            .then(campuses => this.setState({ campuses }));
-
-            // console.log("post delete state",this.state);
+        var campusId=e.target.id;
+        // check if campus has students
+        axios.get(`/api/campuses/${campusId}/students`)
+        .then(students=>{
+            // do not delete if students
+            if(students.data.length){
+                alert('This campus still has students! Move them to a different campus to delete this campus!');
+            }
+            else{
+                axios({
+                    method: 'delete',
+                    url: '/api/campuses/'+campusId
+                })
+                .then(res =>{
+                    axios.get('/api/campuses')
+                        .then(res => res.data)
+                        .then(campuses => this.setState({ campuses }));
+                });
+            }
         });
 
     }
 
     render() {
         const campuses = this.state.campuses;
-        console.log('',campuses);
         var styles = {
             cssFloat:'right'
         };
         return (
             <div>
                 <h1>Campuses</h1>
-                <h2>List of All Campuses (10) <Link to="/campuses/new"><button type="button" className="btn btn-primary" style={styles}>+ Add Campus</button></Link></h2>
+                <h2>List of All Campuses ({campuses.length}) <Link to="/campuses/new"><button type="button" className="btn btn-primary" style={styles}>+ Add Campus</button></Link></h2>
 
 
                 <table className="table table-striped">
